bvh: switched to arrays, camera: adaptive sampling, roulette, vec3: using SIMD

+++ ray.h
@@ -12,7 +12,7 @@ class ray {
     const vec3& direction() const { return dir; }
 
     point3 at(double t) const {
-        return orig + t*dir;
+        return orig + (float)t*dir;
     }
 
   private:
